Refactor Star class to extract separate Marker and CelestialSphereItem classes for clarity
