+++ Assets/Scripts/Core/CelestialSphereItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05b1bae1d27bd4d07a6d9f1afc5f6b36
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
