+++ Assets/Scripts/Core/Marker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b428eec2c0004ccea2bfe97d4f1204c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
